requirements.txt
deepface
numpy
pandas
Pillow
streamlit